libdfp: Fix test-bfp-convertion tests
So test-bfp-convertion convertions assumes that float/double/long double
holds precised values (for instance 12.345) in base 10, where in fact if
the variable is expanded to a more precision value (using prinft or a
multiprecion library) it will hold an different value
(12.345000267028809 in the 12.345 case for float). This patch fixes the
expected value to decimal transformation.